requirements.txt
discord~=2.3.2
pytz~=2023.3
python-dotenv~=1.0.0